+++ src/demo/__init__.py
@@ -1,0 +1,4 @@
+# SPDX-License-Identifier: MIT
+# SPDX-FileCopyrightText: 2023 Berkay Bozkurt <resitberkaybozkurt@gmail.com>
+
+from demos import bdc_demo, db_demo, evp_demo, pipeline_demo
